+++ qiita-client/app/src/main/kotlin/io/github/shofucchi/qiita/ui/feature/common/consts/AppBarTitleConst.kt
@@ -1,0 +1,5 @@
+package io.github.shofucchi.qiita.ui.feature.common.consts
+
+enum class AppBarTitleConst {
+    Article,
+}
